Word2Markdown/requirements.txt
docx2txt==0.8
python_docx==1.1.2
